missing cast to integer added
git-svn-id: https://svn.kflog.org/svn/repos/cumulus/qt4/trunk@3624 c8f415cb-fa42-0410-b21e-89df95f64ae5

+++ cumulus/gpsnmea.cpp
@@ -1171,7 +1171,7 @@ QString GpsNmea::__ExtractConstellation(const QStringList& sentence)
 
       if( ok == true && _lastSatInfo.fixAccuracy != pdop )
         {
-          _lastSatInfo.fixAccuracy = rint(pdop);
+          _lastSatInfo.fixAccuracy = static_cast<int>(rint(pdop));
         }
     }
